Move InDepthNodeVisitor.h back to src/Interpreters.
