+++ src/Interpreters/ConvertStringsToEnumVisitor.h
@@ -4,7 +4,7 @@
 #include <unordered_map>
 #include <unordered_set>
 
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 #include <Parsers/IAST_fwd.h>
