+++ src/Interpreters/FunctionMaskingArgumentCheckVisitor.h
@@ -2,7 +2,7 @@
 
 
 #include <DataTypes/Serializations/ISerialization.h>
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 #include <Parsers/ASTFunction.h>
 #include <Parsers/ASTIdentifier.h>
