+++ src/Interpreters/InterpreterSelectWithUnionQuery.cpp
@@ -21,7 +21,7 @@
 #include <QueryPipeline/QueryPipelineBuilder.h>
 #include <Common/typeid_cast.h>
 
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 
 #include <algorithm>
