+++ src/Interpreters/MarkTableIdentifiersVisitor.h
@@ -2,7 +2,7 @@
 
 #include <Parsers/IAST.h>
 #include <Interpreters/Aliases.h>
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 
 namespace DB
 {
