+++ src/Interpreters/RewriteAnyFunctionVisitor.h
@@ -3,7 +3,7 @@
 #include <unordered_set>
 
 #include <Parsers/IAST.h>
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 
 namespace DB
 {
