+++ src/Interpreters/RewriteCountDistinctVisitor.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <Parsers/IAST.h>
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 #include "Interpreters/TreeRewriter.h"
 
 namespace DB
