+++ src/Interpreters/SubstituteColumnOptimizer.cpp
@@ -7,7 +7,7 @@
 #include <Storages/StorageInMemoryMetadata.h>
 #include <Parsers/ASTSelectQuery.h>
 #include <Poco/Logger.h>
-#include <Parsers/InDepthNodeVisitor.h>
+#include <Interpreters/InDepthNodeVisitor.h>
 #include <Storages/IStorage.h>
